primeiros movimentos das peças - gráfico
